+++ NEWS.md
@@ -7,6 +7,7 @@
  * Added contributor types setting loading for mapping, simple mapping for primary sign and name
  * Added mapping for contributor type free text
  * Added instance-type identifiers settings loading for mapping
+ * Applied caching for external mapping parameters
  
 ## 2019-09-09 v1.6.0
  * Progress mechanism was updated